pyproject.toml
[project]
name = "SmartMessenger"
version = "0.1.0"
description = "ELEKS internship project"
requires-python = ">=3.13,<4.0"
dependencies = [
    "beanie>=1.29.0",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
    "authlib>=1.2.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.0.0",
    "python-multipart >=0.0.6",
    "bcrypt>=4.0.0",
    "itsdangerous >=2.1.2",
    "sqlalchemy-utils>=0.39.0",
    "openai>=1.82.0",
    "asyncpg>=0.30.0",
    "alembic>=1.16.1",
    "fastapi-mail>=1.4.4",
    "pre-commit>=4.2.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"
